Physical review submission prep (#125)
* Slight modification of language in abstract.

Mainly reflecting on one of the reviewers who did not get the point from
the abstract.

* Add the game itself to introduction.

One of the reviewers pointed out that we did not explicitly define some
of the parameters. I'm hoping what I've included does so sufficiently
briefly.

* Minor further rewording to introduction.

* Sign post to a subsection.

* Change obey to satisfy.

* Make very  minor rewording to measure explanation.

* Add a table to validation section.

Shows that chi is recovered as expected.

* Minor modifications to experiments section.

* Increase size of table in experiments section.

* Add one liner saying what skew is.

* Update travis.yml

* Stop testing notebook.

* Compile the pdf.

* Make minor edits suggested by @marcharper.

Thanks Marc! :+1:

* Add three examples of distributions of SSE.

I'm not sure this is necessary, just reflecting on one of the reviewers
what did not know what skew was.

* Simplify large graph and create cp for aux.

* Modify the cover letter.

* Plot skew vs actual score and win prob.

* Remove the regression lines from fixation plots.

* Update text.

- Plot the strategies score/win vs skew(SSE).
- Remove the linear regression